+++ package.json
@@ -40,6 +40,7 @@
     "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
+    "@types/node": "^20.11.7",
     "@types/react-native": "^0.73.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
