+++ src/ui/map-logo.png
